+++ datadog_checks_base/datadog_checks/base/stubs/common.py
@@ -1,10 +1,7 @@
 # (C) Datadog, Inc. 2018-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-from __future__ import division
 
-from collections import namedtuple
+# Module moved to datadog_checks_dev, this import is for backward compatibility
 
-MetricStub = namedtuple('MetricStub', 'name type value tags hostname device')
-ServiceCheckStub = namedtuple('ServiceCheckStub', 'check_id name status tags hostname message')
-HistogramBucketStub = namedtuple('HistogramBucketStub', 'name value lower_bound upper_bound monotonic hostname tags')
+from datadog_checks.dev.stubs.common import *  # noqa: F401, F403
