+++ datadog_checks_base/datadog_checks/base/stubs/log.py
@@ -1,27 +1,7 @@
 # (C) Datadog, Inc. 2019-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-import logging
 
-from ..log import CheckLoggingAdapter as AgentLoggingAdapter
+# Module moved to datadog_checks_dev, this import is for backward compatibility
 
-TRACE_LEVEL = 7
-
-
-class AgentLogger(logging.getLoggerClass()):
-    def trace(self, msg, *args, **kwargs):
-        if self.isEnabledFor(TRACE_LEVEL):
-            self._log(TRACE_LEVEL, msg, args, **kwargs)
-
-    def critical(self, msg, *args, **kwargs):
-        raise NotImplementedError('The critical log level is reserved for agent shutdowns.')
-
-
-class CheckLoggingAdapter(AgentLoggingAdapter):
-    def critical(self, msg, *args, **kwargs):
-        raise NotImplementedError('The critical log level is reserved for agent shutdowns.')
-
-
-def init_logging():
-    logging.addLevelName(TRACE_LEVEL, 'TRACE')
-    logging.setLoggerClass(AgentLogger)
+from datadog_checks.dev.stubs.log import *  # noqa: F401, F403
