+++ datadog_checks_dev/datadog_checks/dev/stubs/__init__.py
@@ -1,0 +1,8 @@
+# (C) Datadog, Inc. 2018-present
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from .aggregator import aggregator
+from .datadog_agent import datadog_agent
+from .tagging import tagger
+
+__all__ = ['aggregator', 'datadog_agent', 'tagger']
